src/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cloudinary==1.29.0
cssutils==2.4.0
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
future==0.18.2
mysql-connector-python==8.0.28
Pillow==9.0.1
protobuf==3.19.4
PyJWT==2.3.0
pynliner==0.8.0
python-environ==0.4.54
pytz==2022.1
six==1.16.0
soupsieve==2.3.2
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.0.0
